% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{combine_directory}
\alias{combine_directory}
\title{Combine school and district directory data}
\usage{
combine_directory(schools, districts)
}
\arguments{
\item{schools}{Processed school directory data}

\item{districts}{Processed district directory data}
}
\value{
Combined data frame
}
\description{
Merges school-level data with district-level data to get a combined
directory with both principal and superintendent information.
}
\keyword{internal}
