% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nddpi_file}
\alias{download_nddpi_file}
\title{Download NDDPI file to temp location}
\usage{
download_nddpi_file(url)
}
\arguments{
\item{url}{URL to download}
}
\value{
Path to downloaded temp file
}
\description{
Download NDDPI file to temp location
}
\keyword{internal}
