% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years from NDDPI data}
\usage{
get_available_years()
}
\value{
Integer vector of available end years
}
\description{
Returns the years available in the NDDPI enrollment data.
Currently 2008-2025 (based on sheet names in the Excel file).
}
\examples{
get_available_years()
}
