% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_graduation.R
\name{get_cache_file_path}
\alias{get_cache_file_path}
\title{Get cache file path for graduation data}
\usage{
get_cache_file_path(end_year, cache_type)
}
\arguments{
\item{end_year}{Academic year end}

\item{cache_type}{Cache type ("grad_tidy" or "grad_wide")}
}
\value{
File path for cache file
}
\description{
Get cache file path for graduation data
}
\keyword{internal}
