% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_district_directory_url}
\alias{get_district_directory_url}
\title{Get directory URL for districts}
\usage{
get_district_directory_url(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024 for 2023-24)}
}
\value{
URL string
}
\description{
Get directory URL for districts
}
\keyword{internal}
