% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_nddpi_url}
\alias{get_nddpi_url}
\title{NDDPI enrollment data URL}
\usage{
get_nddpi_url()
}
\value{
URL string for the enrollment history Excel file
}
\description{
NDDPI enrollment data URL
}
\keyword{internal}
