% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_district_directory}
\alias{get_raw_district_directory}
\title{Download raw district directory data}
\usage{
get_raw_district_directory(end_year)
}
\arguments{
\item{end_year}{School year end (e.g., 2024 for 2023-24)}
}
\value{
Data frame with raw district directory data
}
\description{
Downloads the district directory CSV from ND Insights and parses it.
}
\keyword{internal}
