% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_year}
\alias{parse_school_year}
\title{Convert school year string to end year}
\usage{
parse_school_year(year_str)
}
\arguments{
\item{year_str}{Character string in "YYYY-YY" format}
}
\value{
Integer end year
}
\description{
Converts formats like "2023-24" to 2024 (the end year).
}
\keyword{internal}
