% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_state_enr}
\alias{process_state_enr}
\title{Process state-level enrollment data}
\usage{
process_state_enr(df, end_year)
}
\arguments{
\item{df}{State total row from raw data}

\item{end_year}{School year end}
}
\value{
Processed state row as data frame
}
\description{
Process state-level enrollment data
}
\keyword{internal}
